refactor to use single recent-edits-retriever
