+++ vscode/src/completions/context/retrievers/recent-edits/recent-edits-retriever.ts
@@ -1,5 +1,10 @@
 import { PromptString, contextFiltersProvider } from '@sourcegraph/cody-shared'
+import { ps, psDedent } from '@sourcegraph/cody-shared'
+import type { AutocompleteContextSnippet } from '@sourcegraph/cody-shared'
 import * as vscode from 'vscode'
+import { getLanguageConfig } from '../../../../tree-sitter/language'
+import type { ContextRetriever, ContextRetrieverOptions } from '../../../types'
+import { shouldBeUsedAsContext } from '../../utils'
 
 interface TrackedDocument {
     content: string
@@ -8,18 +13,19 @@ interface TrackedDocument {
     changes: { timestamp: number; change: vscode.TextDocumentContentChangeEvent }[]
 }
 
-export interface DiffAcrossDocuments {
+interface DiffAcrossDocuments {
     diff: PromptString
     uri: vscode.Uri
     languageId: string
     latestChangeTimestamp: number
 }
 
-export class RecentEditsRetriever implements vscode.Disposable {
-    private disposables: vscode.Disposable[] = []
+export class RecentEditsRetriever implements vscode.Disposable, ContextRetriever {
     // We use a map from the document URI to the set of tracked completions inside that document to
     // improve performance of the `onDidChangeTextDocument` event handler.
     private trackedDocuments: Map<string, TrackedDocument> = new Map()
+    public identifier = 'recent-edits'
+    private disposables: vscode.Disposable[] = []
 
     constructor(
         private readonly maxAgeMs: number,
@@ -31,6 +37,31 @@ export class RecentEditsRetriever implements vscode.Disposable {
         this.disposables.push(workspace.onDidChangeTextDocument(this.onDidChangeTextDocument.bind(this)))
         this.disposables.push(workspace.onDidRenameFiles(this.onDidRenameFiles.bind(this)))
         this.disposables.push(workspace.onDidDeleteFiles(this.onDidDeleteFiles.bind(this)))
+    }
+
+    public async retrieve(options: ContextRetrieverOptions): Promise<AutocompleteContextSnippet[]> {
+        const rawDiffs = await this.getDiffAcrossDocuments()
+        const diffs = this.filterCandidateDiffs(rawDiffs, options.document)
+        diffs.sort((a, b) => b.latestChangeTimestamp - a.latestChangeTimestamp)
+
+        const autocompleteContextSnippets = []
+        for (const diff of diffs) {
+            const content = this.getCommentedPromptForCompletions(
+                diff.languageId,
+                diff.uri,
+                diff.diff
+            ).toString()
+            const autocompleteSnippet = {
+                uri: diff.uri,
+                content,
+            } satisfies Omit<AutocompleteContextSnippet, 'startLine' | 'endLine'>
+            autocompleteContextSnippets.push(autocompleteSnippet)
+        }
+        // TODO: add `startLine` and `endLine` to `responses` or explicitly add
+        // remove the startLine and endLine from the response similar to how we do
+        // for BFG.
+        // @ts-ignore
+        return autocompleteContextSnippets.slice(1)
     }
 
     public async getDiffAcrossDocuments(): Promise<DiffAcrossDocuments[]> {
@@ -54,6 +85,36 @@ export class RecentEditsRetriever implements vscode.Disposable {
         const results = await Promise.all(diffPromises)
         diffs.push(...results.filter((result): result is DiffAcrossDocuments => result !== null))
         return diffs
+    }
+
+    public getCommentedPromptForCompletions(
+        languageId: string,
+        filename: vscode.Uri,
+        diff: PromptString
+    ): PromptString {
+        const filePath = PromptString.fromDisplayPath(filename)
+        const languageConfig = getLanguageConfig(languageId)
+        const commentStart = languageConfig ? languageConfig.commentStart : ps`// `
+        const prompt = psDedent`${commentStart} Here is git diff of the recent change made to the file ${filePath} which is used to provide context for the completion:\n${diff}`
+        return prompt
+    }
+
+    public filterCandidateDiffs(
+        allDiffs: DiffAcrossDocuments[],
+        document: vscode.TextDocument
+    ): DiffAcrossDocuments[] {
+        const filterCandidateDiffs: DiffAcrossDocuments[] = []
+        for (const diff of allDiffs) {
+            const currentDocumentLanguageId = document.languageId
+            if (shouldBeUsedAsContext(false, currentDocumentLanguageId, diff.languageId)) {
+                filterCandidateDiffs.push(diff)
+            }
+        }
+        return filterCandidateDiffs
+    }
+
+    public isSupportedForLanguageId(): boolean {
+        return true
     }
 
     public async getDiff(uri: vscode.Uri): Promise<PromptString | null> {
